Add scripts for managing specific dconf settings
With multiple machines using GNOME-based desktops, this really helps
with backing up and restoring specific settings.

In the future I want to be able to pass a list of settings to
backup/restore in via stdin and have the script adapt. Right now it
takes its input from a text file.